sml_server: implement DZG encoding bug workaround
There's an encoding bug in (my) DZG meter(s). It encodes positive values
as if they were with UNSIGNED tag, but then proceeds to tag them as
SIGNED, see the code for an example.

The good thing is that it doesn't attempt to compress negative values,
and we cannot have positive values >= 0x80000000 (that would be >21MW)
and so we can work around this by detecting these values.

Implement this workaround for meters that output their serial number
in the 1-0:96.1.0.255 tag and whose serial number starts "DZG\x00".